Refactoring Chamber of Deputies module to adapt to the new Rosie Refactor
Me and @jtemporal are still planing the refactors to make rosie generic to any dataset.
Refactoring the Chamber of Deputies module to a generic one, is part of it.
